Add files via upload
In Power BI, it is also possible to move your visualization on the reporting canvas by clicking
and then dragging it. You can also switch between different type of charts and visualizations
from the Visualization pane. Power BI attempts to convert your selected fields to the new
visual type as closely as possible.